backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
